Fix initialization bug (repeating enter key #9)
When starting keyd the last key event gets dropped. This happens because
keyd grabs the keyboard before the key up event can get generated. In most
cases this is benign and remedied by hitting the last key that was held
(usually enter) but can cause pathological behaviour in certain libinput
configurations. This patch causes keyd to wait for a neutral keyboard
state before proceeding with initialization.