Added support for hybrid mice/keyboards.
Some keyboards come with a builtin mouse, presently we ignore such
devices altogether. This patch adds support for some of these by passing
EV_REL events through the uinput device unmodified.

It is worth noting that certain touchpads will generate EV_ABS
events which require the device to define the size of
the input area and are consequently unsupported. Fortunately
it seems that most of these devices register as separate input
devices and are ignored by keyd. If this becomes an issue
it may make sense to revisit creating a 1-1 mapping between virtual
and physical devices but this seems unnecessary at present.