Fix layer + overload bug.
Pressing an overloaded key while the layer it is bound to is active
causes permanent activation since it is first cleared by the overloaded
key up event and then subsequently reset by the layer key up event.

In most cases the user is unlikely to have multiple keys corresponding
to the same layer active at the same time, however in the case of
overloaded keys it is sometimes desirable to combine the keycodes to
which they are bound with the same modifier layer to which they are also
bound.

To achieve this we toggle on the overload key down/up rather than
explicitly set the layer to achieve the desired effect in the unusual
case where it is combined with another activation key for the same
layer. This is the same technique we use for layer + layert combos
acting on the same layer.