Fixed accidental layer toggle bug (#44).
This breaks the ability to assign a layer toggle of a given layer
within the layer itself.

E.G

	meta = layer(meta)

	[meta:M]

	a = layert(meta)

No longer toggles the meta layer when meta+a is pressed since the
corresponding meta key up event clears the layer after it is set. Note
that this does not apply to oneshot keys. This configuration seems to be
rare enough to ignore but can be explicitly accounted for with
additional state if necessary.