refactor: reduce memory footprint
Constrain key codes to the range < 256 so they fit inside a single byte.

In theory individual evdev codes may exceed this but in practice most
keys seem to fit in this range, and it generally seems to be more
portable (e.g USB HID spec).

The mod mask has also been demoted to uint8_t since 16 bits was an
unreasonable size to begin with.

This should significantly reduce the memory footprint.