v2.3.0-rc
This is a major release which breaks backward compatibility for
non-trivial configs. (we are still in beta after all :P).

In the absence of too much blowback this will probably become the final
v2 design.

Much of this harkens back to the v1, with some additional simplifications
and enhancements.

See DESIGN.md for a more detailed account.