Fix VT repeat (#227)
Add support for repeat events to the uinput output. This is achieved
using a dedicated thread in the vkbd output logic.

Similar functionality was introduced in 81c3d53f4c before the rewrite. This
solution is a bit cleaner (albeit more expensive) since it confines the
relevant code to the uinput vkbd and more closely simulates behaviour
of the real drivers (e.g by not generating redundant repeat events).