refactor: Major code cleanup
Mostly plumbing.

Summary:

 - Key processing logic remains untouched.
 - Simplified IPC logic.
 - Streamlined the event loop.
 - Consolidated config and descriptor parsing logic into config.[ch].
 - Eliminated internal KEYD_* environment variables in favour of compile time macros.
 - Added color coding to the daemon output.
 - Load all configuration files on initialization.

The last change increases initialization time (on the order of μs) and resident
memory size, but simplifies the config logic and has the benefit of linting all
files immediately while also allowing for the introduction of in-daemon
reloading. This ultimately results in a net saving for the user since
most service managers are horrendously slow :(.

It also means hotplugged devices will use the in-memory config set (which
probably makes more sense anyway) instead of dynamically scanning the config
directory. This is more efficient for small config directories (<100 files),
but less efficient for large ones (uncommon).