Transparently assign right hand modifier keycodes to their corresponding layers (#214)
This seems to be a common point of confusion for new users who don't
understand keyd's modifier semantics. Instead of allowing 'leftshift'
and company to be directly assigned and wreak havoc with properly
bound modifier keys, we now just transparently make modifier keys like
leftshift equivalent to their corresponding layers.

This is technically a breaking change since it makes improperly
bound keys subject to reassignments in the default modifier layer,
but one could argue this was undefined (and rare)
behaviour anyway. A warning is also now issued to ward the
user in the right direction.