swap: Add support for swapping toggled layers (#410)
Sometimes it is desirable to transition directly from one
toggled layer to another.

This commit extends swap() to work not just with keys bound to layer(),
but also toggle() and oneshot(). For swaps which occur one layer
deep (the anticipated use case), this is rougly equivalent of a
clear() followed by toggle(<target>).