keyd-application-mapper: Add Gnome 45 support (#639, #649)
Add support for Gnome 45, which breaks backward extension
compatibility. We now ship and maintain two extensions :/.

The main change in Gnome seems to be the switch from a custom
import implementation to 'standard' ES6 style imports.
The init method also seems to have inexplicably
been swapped out in favour of extending a magic
class.

This patch takes the opportunity to clean up some residual cruft
and move the extension code out of the mapper script.

Specifically:

 - Ship distinct extensions for Gnome 42-44 and Gnome 45 in /usr/local/share/keyd
 - Move the named pipe to XDG_RUNTIME_DIR
 - Initialize the pipe inside the extension instead of the script
   to avoid potential race conditions during initialization.