core: Add trackpad detection
Avoid matching anything which looks like a trackpad on wildcard.

We now also emit a warning that trackpads are unsupported
and likely to break mouse movement.

Many laptops split the trackpad buttons into a separate device node
which is detected as a keyboard, so this does not necessarily
preclude remapping trackpad *buttons*.